delete data with foreign key

+++ app/src/main/java/com/example/ebookapp/Database/DatabaseHandler.java
@@ -7,7 +7,7 @@ import android.util.Log;
 
 public class DatabaseHandler extends SQLiteOpenHelper {
     public final static String DB_NAME = "Library.db";
-    public final static int VERSION = 1;
+    public final static int VERSION = 6;
 
     //table database
     // READER TABLE
@@ -64,15 +64,25 @@ public class DatabaseHandler extends SQLiteOpenHelper {
     }
 
     @Override
+    public void onOpen(SQLiteDatabase db) {
+        super.onOpen(db);
+        if (!db.isReadOnly()) {
+            // Enable foreign key constraints
+            db.execSQL("PRAGMA foreign_keys=ON;");
+        }
+    }
+
+    @Override
     public void onCreate(SQLiteDatabase db) {
         //CreateTable(sqLiteDatabase);
         String TAG = "CREATE TABLE";
         try
         {
-            db.execSQL(CreateBookTB()); // exec book table
-            db.execSQL(CreateReaderTB()); // exec reader table
             db.execSQL(CreateAuthorTB()); // exec author table
             db.execSQL(CreateCategoryTB()); // exec category table
+            db.execSQL(CreateBookTB()); // exec book table
+
+            db.execSQL(CreateReaderTB()); // exec reader table
             db.execSQL(CreateAdminTB()); // exec admin table
             db.execSQL(CreateBorrowingTB()); // exec borrowing
             db.execSQL(CreateBorrowingDetailsTB()); // exec borrowing details
@@ -86,24 +96,31 @@ public class DatabaseHandler extends SQLiteOpenHelper {
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion)
     {
-        db.execSQL("drop table if exists " + BOOK_TB_NAME);
-        db.execSQL("drop table if exists " + READER_TB_NAME);
-        db.execSQL("drop table if exists " + AUTHOR_TB_NAME);
-        db.execSQL("drop table if exists " + ADMIN_TB_NAME);
-        db.execSQL("drop table if exists " + BRW_TB_NAME);
-        db.execSQL("drop table if exists " + BRWDETAILS_TB_NAME);
-        db.execSQL("drop table if exists " + CATEGORY_TB_NAME);
-        onCreate(db);
+        if(oldVersion < newVersion)
+        {
+            db.execSQL("drop table if exists " + BOOK_TB_NAME);
+            db.execSQL("drop table if exists " + READER_TB_NAME);
+            db.execSQL("drop table if exists " + AUTHOR_TB_NAME);
+            db.execSQL("drop table if exists " + ADMIN_TB_NAME);
+            db.execSQL("drop table if exists " + BRW_TB_NAME);
+            db.execSQL("drop table if exists " + BRWDETAILS_TB_NAME);
+            db.execSQL("drop table if exists " + CATEGORY_TB_NAME);
+            onCreate(db);
+        }
+
     }
     private String CreateBookTB()
     {
         String sql = "Create table if not exists " + BOOK_TB_NAME + "(" +
                 "" + BOOK_ID + " integer primary key autoincrement not null," +
                 "" + BOOK_TITLE + " text," +
-                "" + BOOK_AUTHOR_ID + " integer," +
+                "" + BOOK_AUTHOR_ID + " integer not null," +
                 "" + BOOK_YEAR + " date," +
-                "" + BOOK_CATEGORY_ID + " int,"+
-                "" + BOOK_IMAGE + " blob);";
+                "" + BOOK_CATEGORY_ID + " integer not null,"+
+                "" + BOOK_IMAGE + " blob ," +
+                "" + "CONSTRAINT fk_BOOK_CATEGORY_ID FOREIGN KEY (" + BOOK_CATEGORY_ID + ") REFERENCES " + CATEGORY_TB_NAME + "("+ CATEGORY_ID +") ON DELETE CASCADE" + "," +
+                "" + "CONSTRAINT fk_BOOK_AUTHOR_ID FOREIGN KEY (" + BOOK_AUTHOR_ID + ") REFERENCES " + AUTHOR_TB_NAME + "("+ AUTHOR_ID +") ON DELETE CASCADE" +
+                ");";
         return sql;
     }
     private String CreateReaderTB()
@@ -144,18 +161,22 @@ public class DatabaseHandler extends SQLiteOpenHelper {
     {
         String sql = "create table if not exists "+ BRW_TB_NAME +"(" +
                 "" + BRW_ID + " integer primary key autoincrement not null," +
-                "" + BRW_READER_ID + " integer," +
+                "" + BRW_READER_ID + " integer not null," +
                 "" + BRW_BORROW_DAY + " text," +
                 "" + BRW_RETURN_DAY + " text," +
-                "" + BRW_RETURN_TIME + " text);";
+                "" + BRW_RETURN_TIME + " text," + "" +
+                "" + "CONSTRAINT fk_BRW_READER_ID FOREIGN KEY (" + BRW_READER_ID + ") REFERENCES " + READER_TB_NAME + "("+ READER_ID +") ON DELETE CASCADE" +
+                ");";
         return sql;
     }
     private String CreateBorrowingDetailsTB()
     {
         String sql = "create table if not exists " + BRWDETAILS_TB_NAME + "(" +
-                "" + BRWDETAILS_BRW_ID + " integer,"+
-                "" + BRWDETAILS_BOOK_ID + " integer," +
-                " PRIMARY KEY (" +BRWDETAILS_BRW_ID + "," + BRWDETAILS_BOOK_ID + ")" +
+                "" + BRWDETAILS_BRW_ID + " integer not null ,"+
+                "" + BRWDETAILS_BOOK_ID + " integer not null," +
+                " PRIMARY KEY (" +BRWDETAILS_BRW_ID + "," + BRWDETAILS_BOOK_ID + ")," +
+                "" + "FOREIGN KEY (" + BRWDETAILS_BRW_ID + ") REFERENCES " + BRW_TB_NAME + "("+ BRW_ID +") ON DELETE CASCADE" + "," +
+                "" + "FOREIGN KEY (" + BRWDETAILS_BOOK_ID + ") REFERENCES " + BOOK_TB_NAME + "("+ BOOK_ID +") ON DELETE CASCADE" +
                 ");";
 
         return sql;
